backend/requirements.txt
# Flask and Web Framework
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3

# Database
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7  # For PostgreSQL
alembic==1.12.0  # For database migrations

# Security
Werkzeug==2.3.7
PyJWT==2.8.0
bcrypt==4.0.1

# Machine Learning
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
joblib==1.3.2

# Text Processing
nltk==3.8.1
regex==2023.8.8

# Data Visualization (for Jupyter notebook)
matplotlib==3.7.2
seaborn==0.12.2

# Development and Testing
pytest==7.4.2
pytest-flask==1.2.0
python-dotenv==1.0.0

# File handling
Pillow==10.0.0  # For image processing

# Utilities
python-dateutil==2.8.2
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3
Jinja2==3.1.2

# Production server
gunicorn==21.2.0
